move `tcCommands` to new file `latency_emulation.go`
